% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest}
\alias{backtest}
\title{Backtesting via Value-at-Risk (VaR)}
\usage{
backtest(
  x,
  data = NULL,
  window_length = 500,
  p = 0.95,
  portfolio_weights = NULL,
  n.ahead = 1
)
}
\arguments{
\item{x}{An object of class "bekkFit" from the function \link{bekk_fit} or an object of class "bekkSpec".}

\item{data}{A time series if x is of class "bekkSpec".}

\item{window_length}{An integer specifying the length of the rolling window.}

\item{p}{A numerical value that determines the confidence level. The default value is set at 0.99 in accordance with the Basel Regulation.}

\item{portfolio_weights}{A vector determining the portfolio weights to calculate the portfolio VaR. If set to "NULL", the univariate VaR for each series are calculated.}
}
\value{
Returns a S3 class "var" object containing the VaR forecast and respective confidence bands.
}
\description{
Method for calculating VaR from estimated covariance processes (\link{bekk_fit}) or predicted covariances (\link{bekk_forecast}).
}
\examples{
\donttest{

data(StocksBonds)
obj_spec <- bekk_spec()
x1 <- bekk_fit(obj_spec, StocksBonds, QML_t_ratios = FALSE, max_iter = 50, crit = 1e-9)

# single VaRs of series
x2 <- VaR(x1)
plot(x2)

# VaR of equally-weighted portfolio
portfolio_weights <- c(0.5, 0.5)
x3 <- VaR(x1, portfolio_weights = portfolio_weights)
plot(x3)

# VaR of traditional 30/70 weighted bond and stock portfolio
portfolio_weights <- c(0.3, 0.7)
x4 <- VaR(x1, portfolio_weights = portfolio_weights)
plot(x4)

}

}
