% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEKK.R
\name{bekk}
\alias{bekk}
\title{Estimating a BEKK(1, 1) model}
\usage{
bekk(
  r,
  init_values = NULL,
  QML_t_ratios = FALSE,
  seed = NULL,
  max_iter = 50,
  crit = 1e-09
)
}
\arguments{
\item{r}{data input}

\item{init_values}{initial values for BEKK parameter}

\item{QML_t_ratios}{Logical. If QML_t_ratios = 'TRUE', the t-ratios of the BEKK parameter matrices
are exactly calculated via second order derivatives.}

\item{max_iter}{maximum number of BHHH algorithm iterations}

\item{crit}{determiens the precision of the BHHH algorithm}
}
\description{
Estimating a BEKK(1, 1) model
}
\examples{
\donttest{

data(bivariate)
x1 <- bekk(BI, init_values = NULL,
QML_t_ratios = FALSE, max_iter = 50, crit = 1e-9)

summary(x1)

plot(x1)

}
}
