% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_fit.R
\name{bekk_fit}
\alias{bekk_fit}
\title{Estimating BEKK(1, 1) models}
\usage{
bekk_fit(
  spec,
  data,
  QML_t_ratios = FALSE,
  seed = NULL,
  max_iter = 50,
  crit = 1e-09,
  nc = 1
)
}
\arguments{
\item{spec}{object of class "bekkSpec" from function \href{bekk_spec}.}

\item{data}{data input}

\item{QML_t_ratios}{Logical. If QML_t_ratios = 'TRUE', the t-ratios of the BEKK parameter matrices
are exactly calculated via second order derivatives.}

\item{max_iter}{maximum number of BHHH algorithm iterations}

\item{crit}{determiens the precision of the BHHH algorithm}
}
\description{
Estimating BEKK(1, 1) models
}
\examples{
\donttest{

data(bivariate)
x1 <- bekk_fit(BI, init_values = NULL,
QML_t_ratios = FALSE, max_iter = 50, crit = 1e-9)

summary(x1)

plot(x1)

}
}
