% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_forecast.R
\name{bekk_forecast}
\alias{bekk_forecast}
\title{Forecasting conditional volatilities with BEKK models}
\usage{
bekk_forecast(x, n.ahead = 1)
}
\arguments{
\item{x}{A fitted bekk model of class bekk from the \link{bekk} function}

\item{n.ahead}{Number of periods to forecast conditional volatility. Default is a one-period ahead forecast.}
}
\description{
Forecasting conditional volatilities with BEKK models
}
\examples{
\donttest{

data(StocskBonds)
obj_spec <- bekk_spec()
x1 <- bekk_fit(obj_spec, StocksBonds, QML_t_ratios = FALSE, max_iter = 50, crit = 1e-9)

summary(x1)

plot(x1)

}
}
