% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_spec.R
\name{bekk_spec}
\alias{bekk_spec}
\title{BEKK specification method}
\usage{
bekk_spec(
  model = list(type = "bekk", asymmetric = FALSE),
  init_values = NULL,
  signs = NULL,
  N = NULL
)
}
\arguments{
\item{model}{List containing the model type specification: Currently implemented are "bekk", "diagBekk", "scalarBekk". Moreover it can be specified whether the model should be estimated allowing for asymmetric volatility structure.}

\item{init_values}{initial values for BEKK parameter during BHHH algorithm. If object from this function is passed to \link{bekk_sim}, init_values are used as parameters for data generating process.}

\item{signs}{Vector specifying asymmetry.}

\item{Integer}{specifying the dimension of the BEKK model. Only relevant for \link{bekk_sim}.}
}
\description{
BEKK specification method
}
