% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_virf.R
\name{bekk_virf}
\alias{bekk_virf}
\title{Estimating multivariate volatility impulse response functions (VIRF) for BEKK models}
\usage{
bekk_virf(fit, time = 1, q = c(0.05, 0.5, 0.95), periods = 10)
}
\arguments{
\item{fit}{An object of class "bekkfit" from function \link{bekk_fit}.}

\item{time}{Time instace to calculate VIRFs for.}

\item{q}{Quantiles for VIRFs.}

\item{periods}{Periods ahead for VIRFs.}
}
\value{
Returns an object of class "bekkVIRF".
}
\description{
Method for estimating VIRFs of N-dimensional BEKK models.
}
