% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEKKs.R
\docType{package}
\name{BEKKs}
\alias{BEKKs}
\title{BEKKs: Volatility modelling}
\description{
This package implements MGARCH estimation techniques for conditional volatility modelling.\cr
}
\details{
The main functions to retrieve structural impact matrices are:
\itemize{
\item \tabular{ll}{ \code{\link{bekk}} \tab Estimates a BEKK(1,1,1) model,}
}
}
\author{
\itemize{
\item Markus Fülle  \email{fuelle@uni-goettingen.de}
\item Helmut Herwartz \email{hherwartz@uni-goettingen.de}
\item Alexander Lange \email{alexander.lange@uni-goettingen.de}
}
}
