% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch}
\alias{garch}
\title{Estimating an GARCH(p, q) model}
\usage{
garch(r, p, q, max_iter = 3e+05, crit = 1e-12)
}
\arguments{
\item{r}{data input}

\item{p}{GARCH order}

\item{q}{ARCH order}

\item{max_iter}{maximum number of BHHH algorithm iterations}

\item{crit}{determiens the precision of the BHHH algorithm}
}
\description{
Estimating an GARCH(p, q) model
}
