% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.backtest.R
\name{plot.backtest}
\alias{plot.backtest}
\title{backtest method}
\usage{
\method{plot}{backtest}(x, ...)
}
\arguments{
\item{x}{An object of class "backtest" from function \link{backtest}.}

\item{...}{Further arguments to be passed to and from other methods.}
}
\description{
Generic 'backtest' method. More details on 'backtest' are described in \link{backtest}
}
