% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_fit.R
\name{bekk_fit}
\alias{bekk_fit}
\title{Estimating multivariate BEKK-type volatility models}
\usage{
bekk_fit(
  spec,
  data,
  QML_t_ratios = FALSE,
  seed = NULL,
  max_iter = 50,
  crit = 1e-09
)
}
\arguments{
\item{spec}{An object of class "bekkSpec" from function \link{bekk_spec}.}

\item{data}{A multivariate data object. Can be a numeric matrix or ts/xts/zoo object.}

\item{QML_t_ratios}{Logical. If QML_t_ratios = 'TRUE', the t-ratios of the BEKK parameter matrices
are exactly calculated via second order derivatives.}

\item{max_iter}{Maximum number of BHHH algorithm iterations.}

\item{crit}{Determiens the precision of the BHHH algorithm.}
}
\description{
Method for fitting a variety of N-dimensional BEKK models.
}
\details{
The BEKK optimization routine is based on the Berndt–Hall–Hall–Hausman (BHHH) algorithm and is inspired by the study of Hafner and Herwartz (2008).
The authors provide analytical formulas for the score and Hessian of several MGARCH models in a QML framework and show that analytical derivations significantly outperform numerical methods.
}
\examples{
\donttest{

data(StocksBonds)
obj_spec <- bekk_spec()
x1 <- bekk_fit(obj_spec, StocksBonds, QML_t_ratios = FALSE, max_iter = 50, crit = 1e-9)

summary(x1)

plot(x1)

}
}
\references{
Hafner and Herwartz (2005). Analytical quasi maximum likelihood inference in multivariate volatility models. Metrika, 67, 219-239.
}
