% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_forecast.R
\name{bekk_forecast}
\alias{bekk_forecast}
\title{Forecasting conditional volatility with BEKK(1,1) models}
\usage{
bekk_forecast(spec, forecast_length = 1)
}
\arguments{
\item{spec}{A fitted bekk model of class bekk from the \link{bekk} function}

\item{forecast_length}{Number of periods to forecast conditional volatility. Default is a one-period ahead forecast.}
}
\description{
Forecasting conditional volatility with BEKK(1,1) models
}
\examples{
\donttest{

data(bivariate)
x1 <- bekk_fit(BI, init_values = NULL,
QML_t_ratios = FALSE, max_iter = 50, crit = 1e-9)

summary(x1)

plot(x1)

}
}
