% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_sim.R
\name{bekk_sim}
\alias{bekk_sim}
\title{Simulating BEKK models}
\usage{
bekk_sim(spec, nobs)
}
\arguments{
\item{spec}{A spec object of class "bekkSpec" from the function \link{bekk_spec} or a fitted bekk model of class "bekkFit" from the \link{bekk_fit} function}

\item{nobs}{Number of observations of the simulated sample}
}
\description{
Method for simulating a N-dimensional BEKK model.
}
