% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekk_spec.R
\name{bekk_spec}
\alias{bekk_spec}
\title{BEKK specification method}
\usage{
bekk_spec(
  model = list(type = "bekk", asymmetric = FALSE),
  init_values = NULL,
  signs = NULL,
  N = NULL,
  compare = FALSE
)
}
\arguments{
\item{model}{List containing the model type specification: Currently implemented are "bekk", "dbekk", "sbekk".
Moreover it can be specified whether the model should be estimated allowing for asymmetric volatility structure.}

\item{init_values}{initial values for \href{bekk_fit} during BHHH algorithm. It can be either a numerical vector of suitable dimension, or a character vector i.e. "random" to use a random starting values generator, or
"simple" for relying on a simple initial values generator based on typical values for BEKK parameter found in the literature. If object from this function is passed to \link{bekk_sim}, init_values are used as parameters for data generating process.}

\item{signs}{Vector specifying asymmetry.}

\item{N}{Integer specifying the dimension of the BEKK model. Only relevant for \link{bekk_sim}.}
}
\description{
BEKK specification method
}
