% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEKKs.R
\docType{package}
\name{BEKKs}
\alias{BEKKs}
\title{BEKKs: Volatility modelling}
\description{
BEKKs: Volatility modelling
}
\author{
\itemize{
\item Alexander Lange \email{alexander.lange@uni-goettingen.de}
\item Helmut Herwartz \email{hherwartz@uni-goettingen.de}
}
}
