% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch.R
\name{garch}
\alias{garch}
\title{Estimating an ARCH(q) model}
\usage{
garch(r, p, q, max_iter = 30000, crit = 1e-09)
}
\arguments{
\item{r}{data input}

\item{p}{GARCH order}

\item{q}{ARCH order}

\item{max_iter}{maximum number of BHHH algorithm iterations}

\item{crit}{determiens the precision of the BHHH algorithm}
}
\description{
Estimating an ARCH(q) model
}
