% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virf.R
\name{virf}
\alias{virf}
\title{Estimating multivariate volatility impulse response functions (VIRF) for BEKK models}
\usage{
virf(x, time = 1, q = 0.05, index_series = 1, n.ahead = 10, ci = 0.9)
}
\arguments{
\item{x}{An object of class "bekkfit" from function \link{bekk_fit}.}

\item{time}{Time instace to calculate VIRFs for.}

\item{q}{A vector specifying the quantiles to be considered for a shock on which basis the VIRFs are generated.}

\item{n.ahead}{An integer defining the number periods for which the VIRFs are generated.}
}
\value{
Returns an object of class "bekkVIRF".
}
\description{
Method for estimating VIRFs of N-dimensional BEKK models. Currently, only VIRFs for symmetric BEKK models are implemented.
}
\examples{
\donttest{

data(StocksBonds)
obj_spec <- bekk_spec()
x1 <- bekk_fit(obj_spec, StocksBonds, QML_t_ratios = FALSE, max_iter = 50, crit = 1e-9)

# 20 day ahead VIRFs and 90\% CI for a Shock in the 5\% quantile  of Bonds (i.e. series=2)
# shock is supposed to occur at day 50
x2 <- virf(x1, time = 50, q = 0.05, index_series=2, n.ahead = 20, ci = 0.90)
plot(x2)
}
}
